ray cast non operationnel

+++ src/libft/libft.h
@@ -23,7 +23,7 @@ typedef struct s_list
 	struct s_list	*next;
 }	t_list;
 
-int					ft_abs(int x);
+int					ft_abs(float x);
 
 char				**ft_realloc_one(char **old, int len, char *new_line);
